rename server scripts
