+++ packages/language-server/lib/cancellationToken.ts
