+++ packages/language-server/lib/config.ts
