+++ packages/language-server/lib/project/typescriptProjectProvider.ts
@@ -3,9 +3,9 @@ import * as path from 'path-browserify';
 import type * as ts from 'typescript/lib/tsserverlibrary';
 import * as vscode from 'vscode-languageserver';
 import { URI } from 'vscode-uri';
-import { getInferredCompilerOptions } from '../common/utils/inferredCompilerOptions';
-import { isFileInDir } from '../common/utils/isFileInDir';
-import { createUriMap } from '../common/utils/uriMap';
+import { getInferredCompilerOptions } from './inferredCompilerOptions';
+import { isFileInDir } from '../utils/isFileInDir';
+import { createUriMap } from '../utils/uriMap';
 import { ServerMode, ServerProjectProvider, TypeScriptServerPlugin } from '../types';
 import { TypeScriptServerProject, createTypeScriptServerProject } from './typescriptProject';
 import { WorkspacesContext } from './basicProjectProvider';
