+++ packages/language-server/lib/register/registerLanguageFeatures.ts
@@ -1,17 +1,17 @@
 import * as embedded from '@volar/language-service';
 import * as vscode from 'vscode-languageserver';
-import { AutoInsertRequest, FindFileReferenceRequest } from '../../../protocol';
-import { ServerRuntimeEnvironment, InitializationOptions, ServerMode, ServerProjectProvider } from '../../types';
-import { createDocuments } from '../documents';
+import { AutoInsertRequest, FindFileReferenceRequest } from '../../protocol';
+import { ServerRuntimeEnvironment, InitializationOptions, ServerMode, ServerProjectProvider } from '../types';
+import { createDocumentManager } from '../documentManager';
 
-export function register(
+export function registerLanguageFeatures(
 	connection: vscode.Connection,
 	projectProvider: ServerProjectProvider,
 	initParams: vscode.InitializeParams,
 	initOptions: InitializationOptions,
 	semanticTokensLegend: vscode.SemanticTokensLegend,
 	runtime: ServerRuntimeEnvironment,
-	documents: ReturnType<typeof createDocuments>,
+	documents: ReturnType<typeof createDocumentManager>,
 ) {
 
 	let lastCompleteUri: string;
