+++ packages/language-server/lib/schemaRequestHandlers/http.ts
