+++ packages/language-server/lib/utils/isFileInDir.ts
