+++ packages/language-server/lib/utils/uriMap.ts
