+++ packages/language-server/lib/workspaceFolderManager.ts
